examples now use their own ollama `generate`
